+++ Src/FluentAssertions/EventRaisingExtensions.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using FluentAssertions.Common;
@@ -13,84 +14,139 @@ namespace FluentAssertions
     public static class EventRaisingExtensions
     {
         /// <summary>
-        /// Asserts that all occurrences of the event originated from the <param name="expectedSender"/>.
+        /// Asserts that all occurrences of the event originated from the <param name="expectedSender"/> and
+        /// returns only the events that came from that sender.
         /// </summary>
-        public static IEventRecorder WithSender(this IEventRecorder eventRecorder, object expectedSender)
+        public static IEventRecording WithSender(this IEventRecording eventRecording, object expectedSender)
         {
-            foreach (RecordedEvent recordedEvent in eventRecorder)
-            {
-                if (!recordedEvent.Parameters.Any())
-                {
-                    throw new ArgumentException(string.Format(
-                        "Expected event from sender <{0}>, but event {1} does not include any arguments",
-                        expectedSender, eventRecorder.EventName));
-                }
+            var eventsForSender = new List<OccurredEvent>();
+            var otherSenders = new List<object>();
 
-                object actualSender = recordedEvent.Parameters.First();
-                Execute.Assertion
-                    .ForCondition(ReferenceEquals(actualSender, expectedSender))
-                    .FailWith("Expected sender {0}, but found {1}.", expectedSender, actualSender);
+            foreach (OccurredEvent @event in eventRecording)
+            {
+                bool hasSender = Execute.Assertion
+                    .ForCondition(@event.Parameters.Any())
+                    .FailWith("Expected event from sender {0}, " +
+                              $"but event {eventRecording.EventName} does not have any parameters", expectedSender);
+
+                if (hasSender)
+                {
+                    object sender = @event.Parameters.First();
+                    if (ReferenceEquals(sender, expectedSender))
+                    {
+                        eventsForSender.Add(@event);
+                    }
+                    else
+                    {
+                        otherSenders.Add(sender);
+                    }
+                }
             }
 
-            return eventRecorder;
+            Execute.Assertion
+                .ForCondition(eventsForSender.Any())
+                .FailWith("Expected sender {0}, but found {1}.",
+                    () => expectedSender,
+                    () => otherSenders.Distinct());
+
+            return new FilteredEventRecording(eventRecording, eventsForSender);
         }
 
         /// <summary>
-        /// Asserts that at least one occurrence of the event had at least one of the arguments matching a predicate.
+        /// Asserts that at least one occurrence of the events had at least one of the arguments matching a predicate. Returns
+        /// only the events that matched that predicate.
         /// </summary>
-        public static IEventRecorder WithArgs<T>(this IEventRecorder eventRecorder, Expression<Func<T, bool>> predicate)
+        public static IEventRecording WithArgs<T>(this IEventRecording eventRecording, Expression<Func<T, bool>> predicate)
         {
             Guard.ThrowIfArgumentIsNull(predicate, nameof(predicate));
 
             Func<T, bool> compiledPredicate = predicate.Compile();
 
-            if (!eventRecorder.First().Parameters.OfType<T>().Any())
+            bool hasArgumentOfRightType = false;
+            var eventsMatchingPredicate = new List<OccurredEvent>();
+
+            foreach (OccurredEvent @event in eventRecording)
             {
-                throw new ArgumentException("No argument of event " + eventRecorder.EventName + " is of type <" + typeof(T) + ">.");
+                var typedParameters = @event.Parameters.OfType<T>().ToArray();
+                if (typedParameters.Any())
+                {
+                    hasArgumentOfRightType = true;
+                }
+
+                if (typedParameters.Any(parameter => compiledPredicate(parameter)))
+                {
+                    eventsMatchingPredicate.Add(@event);
+                }
             }
 
-            if (eventRecorder.All(recordedEvent => !recordedEvent.Parameters.OfType<T>().Any(parameter => compiledPredicate(parameter))))
+            if (!hasArgumentOfRightType)
+            {
+                throw new ArgumentException("No argument of event " + eventRecording.EventName + " is of type <" + typeof(T) + ">.");
+            }
+
+            if (!eventsMatchingPredicate.Any())
             {
                 Execute.Assertion
                     .FailWith("Expected at least one event with arguments matching {0}, but found none.", predicate.Body);
             }
 
-            return eventRecorder;
+            return new FilteredEventRecording(eventRecording, eventsMatchingPredicate);
         }
 
         /// <summary>
-        /// Asserts that at least one occurrence of the event had arguments matching all predicates.
+        /// Asserts that at least one of the occurred events has arguments the match the predicates in the same order. Returns
+        /// only the events that matched those predicates.
         /// </summary>
-        public static IEventRecorder WithArgs<T>(this IEventRecorder eventRecorder, params Expression<Func<T, bool>>[] predicates)
+        /// <remarks>
+        /// If a <c>null</c> is provided as predicate argument, the corresponding event parameter value is ignored.
+        /// </remarks>
+        public static IEventRecording WithArgs<T>(this IEventRecording eventRecording, params Expression<Func<T, bool>>[] predicates)
         {
             Func<T, bool>[] compiledPredicates = predicates.Select(p => p?.Compile()).ToArray();
 
-            if (!eventRecorder.First().Parameters.OfType<T>().Any())
-            {
-                throw new ArgumentException("No argument of event " + eventRecorder.EventName + " is of type <" + typeof(T) + ">.");
-            }
+            bool hasArgumentOfRightType = false;
+            var eventsMatchingPredicate = new List<OccurredEvent>();
 
-            bool expected = eventRecorder.Any(recordedEvent =>
+            foreach (OccurredEvent @event in eventRecording)
             {
-                T[] parameters = recordedEvent.Parameters.OfType<T>().ToArray();
-                int parametersToCheck = Math.Min(parameters.Length, predicates.Length);
-
-                bool isMatch = true;
-                for (int i = 0; i < parametersToCheck && isMatch; i++)
+                var typedParameters = @event.Parameters.OfType<T>().ToArray();
+                if (typedParameters.Any())
                 {
-                    isMatch = compiledPredicates[i]?.Invoke(parameters[i]) ?? true;
+                    hasArgumentOfRightType = true;
                 }
 
-                return isMatch;
-            });
+                if (predicates.Length > typedParameters.Length)
+                {
+                    throw new ArgumentException(
+                        $"Expected the event to have at least {predicates.Length} parameters of type {typeof(T)}, but only found {typedParameters.Length}.");
+                }
 
-            if (!expected)
-            {
-                Execute.Assertion
-                    .FailWith("Expected at least one event with arguments matching {0}, but found none.", string.Join(" | ", predicates.Where(p => p != null).Select(p => p.Body.ToString())));
+                bool isMatch = true;
+                for (int index = 0; index < predicates.Length && isMatch; index++)
+                {
+                    isMatch = compiledPredicates[index]?.Invoke(typedParameters[index]) ?? true;
+                }
+
+                if (isMatch)
+                {
+                    eventsMatchingPredicate.Add(@event);
+                }
             }
 
-            return eventRecorder;
+            if (!hasArgumentOfRightType)
+            {
+                throw new ArgumentException("No argument of event " + eventRecording.EventName + " is of type <" + typeof(T) + ">.");
+            }
+
+            if (!eventsMatchingPredicate.Any())
+            {
+                Execute
+                    .Assertion
+                    .FailWith("Expected at least one event with arguments matching {0}, but found none.",
+                    string.Join(" | ", predicates.Where(p => p != null).Select(p => p.Body.ToString())));
+            }
+
+            return new FilteredEventRecording(eventRecording, eventsMatchingPredicate);
         }
     }
 }